Handler loading code.
Moved YouTube handler to its own file.
Now every handler should live in the handlers
directory (with a good name). Each handler will
be a function which returns the handler itself.
The register code of VideoHandlers will invoke
that function and store the handler in the
handlers array.
The video tag creation have been moved to
VideoHandlers object, and every handle can use
it passing the DOM object to substitute, the
video URL and some options.

Thanks to Yaoha Chen, for providing some ideas
implemented in this commit.