API change for handlers that need the DOM node.
This will allow coding for video players like
JWPlayer.

The API expects that this function will return a
null value when the handler don't know how to
handle the URL or the DOM node. Otherwise it will
invoke start() in the object returned by this
function. This allow to be lean in case we only
need to test for the URL, but also create and
reuse some of the tests for later processing if
needed.